shrinking data types to save space

+++ imspy_connector/src/py_simulation.rs
@@ -15,11 +15,11 @@ impl PyTimsTofSyntheticsDDA {
         PyTimsTofSyntheticsDDA { inner: TimsTofSyntheticsDDA::new(path).unwrap() }
     }
 
-    pub fn build_frame(&self, frame_id: i64) -> PyTimsFrame {
+    pub fn build_frame(&self, frame_id: u32) -> PyTimsFrame {
         PyTimsFrame { inner: self.inner.build_frame(frame_id) }
     }
 
-    pub fn build_frames(&self, frame_ids: Vec<i64>, num_threads: usize) -> Vec<PyTimsFrame> {
+    pub fn build_frames(&self, frame_ids: Vec<u32>, num_threads: usize) -> Vec<PyTimsFrame> {
         let frames = self.inner.build_frames(frame_ids, num_threads);
         frames.iter().map(|x| PyTimsFrame { inner: x.clone() }).collect::<Vec<_>>()
     }
